Rename args folder
